Refactoring of note editing. More work on Yaeltex Turn.

+++ src/main/java/de/mossgrabers/framework/controller/valuechanger/IValueChanger.java
@@ -158,12 +158,13 @@ public interface IValueChanger
 
 
     /**
-     * Translate the value in the range from [0, max-1] to [0.0, 1.0].
+     * Translate the value in the range from [0, max-1] to [0.0, 1.0]. Allows doubles as input for
+     * calculating small offset values.
      *
      * @param value The DAW value
      * @return The MIDI value
      */
-    double toNormalizedValue (int value);
+    double toNormalizedValue (double value);
 
 
     /**
